requirements.txt
chardet==5.2.0
datasets==3.6.0
Flask==3.1.0
gunicorn==23.0.0
h5py==3.13.0
numpy==1.26.2
Pillow==11.2.1
requests==2.32.3
tensorflow==2.19.0
